Move deviceadmin tasks into tasks.py.
